+++ nitrocli/src/tests/mod.rs
@@ -85,6 +85,15 @@ impl Builder {
     self
   }
 
+  /// Set the `PATH` used for looking up extensions.
+  fn path<P>(mut self, path: P) -> Self
+  where
+    P: Into<ffi::OsString>,
+  {
+    self.0.path = Some(path.into());
+    self
+  }
+
   /// Set the password to use for certain operations.
   fn password<P>(mut self, password: P) -> Self
   where
@@ -102,6 +111,7 @@ impl Builder {
 
 struct Nitrocli {
   model: Option<nitrokey::Model>,
+  path: Option<ffi::OsString>,
   admin_pin: Option<ffi::OsString>,
   user_pin: Option<ffi::OsString>,
   new_admin_pin: Option<ffi::OsString>,
@@ -113,6 +123,7 @@ impl Nitrocli {
   pub fn new() -> Self {
     Self {
       model: None,
+      path: None,
       admin_pin: Some(nitrokey::DEFAULT_ADMIN_PIN.into()),
       user_pin: Some(nitrokey::DEFAULT_USER_PIN.into()),
       new_admin_pin: None,
@@ -166,6 +177,7 @@ impl Nitrocli {
     let ctx = &mut crate::RunCtx {
       stdout: &mut stdout,
       stderr: &mut stderr,
+      path: self.path.clone(),
       admin_pin: self.admin_pin.clone(),
       user_pin: self.user_pin.clone(),
       new_admin_pin: self.new_admin_pin.clone(),